Move frontend interactivity tests to own file
